Ivana's milestone 6 (in progress). - Audio passthrough to mixer. - Audible static, have to check hardware block timing.

+++ sdk/audioTest/src/audioCodecCom.h
@@ -22,7 +22,7 @@ void dataIn(int samplesToRead, volatile u64* toBuffer, int offset);
 // samplesToSend:	The number of samples to send.
 // fromBuffer:		A pointer to the buffer containing the samples to send.
 // offset:			Start sending samples from the specified offset of the buffer above.
-void dataOut(int samplesToSend, volatile u64* fromBuffer, int offset, bool circularBufferOnly);
+void dataOut(int samplesToSend, volatile u32* fromBuffer, int offset, bool circularBufferOnly);
 
 void equalize();
